Removed unused tint property from CategoryFIlter
