+++ QuizApp/Domain/Quiz/Models/Category.swift
@@ -4,4 +4,13 @@ enum Category: String, CaseIterable {
     case sport = "SPORT"
     case movies = "MOVIES"
 
+    var named: String {
+        switch self {
+        case .uncategorized:
+            return LocalizedStrings.all.localizedString
+        default:
+            return self.rawValue
+        }
+    }
+
 }
