+++ QuizApp/Presentation/Home/Models/Category+Color.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+extension Category {
+
+    var color: Color {
+        switch self {
+        case .sport:
+            return .accentYellow
+        case .movies:
+            return .accentBlue
+        default:
+            return .white
+        }
+    }
+
+}
