+++ QuizApp/Presentation/Home/Models/Category+UIColor.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+extension Category {
+
+    var uiColor: UIColor {
+        switch self {
+        case .sport:
+            return .accentYellow
+        case .movies:
+            return .accentBlue
+        default:
+            return .white
+        }
+    }
+
+}
