+++ QuizApp/Presentation/Home/Models/CategoryFilter.swift
@@ -5,7 +5,6 @@ struct CategoryFilter: Hashable {
     let index: Int
     let title: String
     let category: Category
-    let tint: UIColor
 
     func hash(into hasher: inout Hasher) {
         hasher.combine(index)
